pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibe-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "mcp>=1.26.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]

[project.scripts]
vibe-flow = "vibe_flow.cli:main"

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"scratch/**" = ["TD", "FIX", "INP001", "T201"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
